Added C:\OpenSrc\ctags_svn\Test\bug1950327.js
In some cases methods were not parsing the ending
terminator properly resulting in missed tags.

When creating new functions and classes the parser
was not taking into account the scope of the token.
This resulted in some functions and classes from
being ignored even though they were for a different
class.



git-svn-id: https://svn.code.sf.net/p/ctags/code/trunk@664 c5d04d22-be80-434c-894e-aa346cc9e8e8